+++ gtk/src/windowreferences.cpp
@@ -603,7 +603,7 @@ void WindowReferences::html_write_action_bar (HtmlWriter2& htmlwriter, bool topb
       htmlwriter.hyperlink_add ("prev", "[prev]");
       htmlwriter.text_add (" ");
     }
-    htmlwriter.text_add ("Items " + convert_to_string (lower_boundary + 1) + " - " + convert_to_string (upper_boundary) + " of " + convert_to_string (references.size()));
+    htmlwriter.text_add ("Items " + convert_to_string ((unsigned int)(lower_boundary + 1)) + " - " + convert_to_string ((unsigned int)upper_boundary) + " of " + convert_to_string ((unsigned int)references.size()));
     if (upper_boundary < references.size()) {
       htmlwriter.text_add (" ");
       htmlwriter.hyperlink_add ("next", "[next]");
@@ -638,7 +638,7 @@ void WindowReferences::html_write_action_page (HtmlWriter2& htmlwriter)
   // If there are any references at all, offer the option to dismiss the whole lot.
   if (!references.empty()) {
     htmlwriter.paragraph_open ();
-    htmlwriter.hyperlink_add ("dismiss all", "Dismiss the whole lot of " + convert_to_string (references.size()) + " references");
+    htmlwriter.hyperlink_add ("dismiss all", "Dismiss the whole lot of " + convert_to_string ((unsigned int)references.size()) + " references");
     htmlwriter.paragraph_close ();
   }  
   // If a reference has been clicked, offer the option to hide it from now on.
